requirements.txt
asgiref==3.5.0
Django==4.0.1
docx-mailmerge==0.5.0
lxml==4.7.1
objects==0.3.1
psycopg2-binary==2.9.3
python-docx==0.8.11
pytz==2021.3
sqlparse==0.4.2
